refactor: change repository path domain to infra and pickup to external folder
