+++ src/main/java/kitchenpos/global/external/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.global.infra;
+package kitchenpos.global.external;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
