+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -2,10 +2,10 @@ package kitchenpos.product.application;
 
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.infra.MenuRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
-import kitchenpos.global.infra.PurgomalumClient;
+import kitchenpos.product.infra.ProductRepository;
+import kitchenpos.global.external.PurgomalumClient;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
