+++ src/main/java/kitchenpos/product/infra/JpaProductRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.product.domain;
+package kitchenpos.product.infra;
 
+import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
